# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Signals DPG API",
    "description": "Network-aware Signals DPG API — items, actions, events, consent, network fetch, admin.\n\nUnless marked otherwise, operations require authentication via either the `apiKeyAuth` or `sessionAuth` scheme (the spec default). Public operations carry no Authorizations section; the two inter-instance `*_local` operations use the service-to-service `peerAuth` scheme instead. Admin and aggregator operations additionally require the `x-acting-org-id` header (optional on action operations). See `docs/operations/integrating-dpgs.md` for the full auth model.",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Service API key used by integrating DPGs (aggregator-dpg, voice-dpg) and other machine clients (apps/api/plugins/auth/auth_middleware.ts). Takes priority over session auth: if present and invalid the request is rejected outright, with no fallback to a session. Routes under /api/v1/admin and /api/v1/aggregator additionally require an `x-acting-org-id` header identifying the organization the request acts on behalf of — see docs/operations/integrating-dpgs.md."
      },
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Browser session cookie issued by better-auth after sign-in, used by the web UI (apps/api/plugins/auth/auth_middleware.ts). Checked as a fallback only when x-api-key is absent."
      },
      "peerAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-instance-token",
        "description": "Inter-instance HMAC token (paired with an `x-instance-timestamp` header), used only by peer Signals instances for the network `*_local` operations (src/middleware/peer_instance_guard.ts). Not for external callers."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/admin/aggregator/upsert": {
      "post": {
        "tags": [
          "admin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "external_id": {
                    "type": "string",
                    "minLength": 1,
                    "description": "aggregator-dpg primary key for this aggregator"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "description": "display name"
                  },
                  "slug": {
                    "type": "string",
                    "minLength": 1,
                    "pattern": "^[a-z0-9-]+$",
                    "description": "stable url-safe identifier; lookup key for upsert"
                  },
                  "logo_url": {
                    "description": "optional logo url shown in dashboards",
                    "type": "string",
                    "format": "uri"
                  },
                  "domains": {
                    "description": "item domains this aggregator's dashboard reports on (e.g. ['seeker'] or ['seeker','provider']). Persisted as org.metadata.domains. Defaults to [] when omitted; dashboard returns 400 NO_DOMAINS_CONFIGURED until set.",
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "metadata": {
                    "description": "opaque metadata stored alongside external_id on the org",
                    "type": "object",
                    "additionalProperties": {}
                  }
                },
                "required": [
                  "external_id",
                  "name",
                  "slug"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-acting-org-id",
            "required": true,
            "description": "Organization this request acts on behalf of. Required for admin/aggregator operations."
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "org_id": {
                      "type": "string",
                      "description": "Signals organization id (prefixed org_<uuid>)"
                    },
                    "created": {
                      "type": "boolean",
                      "description": "true on first upsert; false when an existing org was updated"
                    }
                  },
                  "required": [
                    "org_id",
                    "created"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/aggregator/default": {
      "post": {
        "tags": [
          "admin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "org_id": {
                    "type": "string",
                    "minLength": 1,
                    "description": "Signals organization id (org_<uuid>) that becomes the default; must be type=aggregator"
                  },
                  "bindings": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^[a-z0-9_]+\\/[a-z0-9_]+$"
                    },
                    "description": "served-domain bindings this org becomes the default for, e.g. ['blue_dot/seeker','blue_dot/provider']. Replaces the org's current set; [] clears it."
                  }
                },
                "required": [
                  "org_id",
                  "bindings"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-acting-org-id",
            "required": true,
            "description": "Organization this request acts on behalf of. Required for admin/aggregator operations."
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "org_id": {
                      "type": "string"
                    },
                    "bindings": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "the org's bindings after the write"
                    },
                    "cleared_from": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "org_id": {
                            "type": "string"
                          },
                          "binding": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "org_id",
                          "binding"
                        ],
                        "additionalProperties": false
                      },
                      "description": "bindings taken off another org by this call"
                    }
                  },
                  "required": [
                    "org_id",
                    "bindings",
                    "cleared_from"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://dpg.example.com",
      "description": "Your deployment's public host (set per network instance)"
    },
    {
      "url": "http://localhost:2742",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "apiKeyAuth": []
    },
    {
      "sessionAuth": []
    }
  ],
...
}